XIVY-18805 create dedicated ai directory
Move AI related code to a dedicated directory.
