+++ extension/src/index.ts
@@ -1,19 +1,18 @@
 import 'reflect-metadata';
 import type { ExtensionContext } from 'vscode';
 import { Messenger, type MessengerDiagnostic } from 'vscode-messenger';
+import { registerTools } from './ai/tools/tools';
 import { registerCommand } from './base/commands';
 import { config } from './base/configurations';
 import { validateAndSyncJavaVersion } from './base/java-version-validation';
 import { askToReloadWindow } from './base/reload-window';
 import { setStatusBarIcon, showStatusBarQuickPick } from './base/status-bar';
-import { registerTool } from './base/tools';
 import { addDevContainer } from './dev-container/command';
 import { conditionalWelcomePage, showWelcomePage } from './editors/welcome-page/welcome-page';
 import { IvyDiagnostics } from './engine/diagnostics';
 import { IvyEngineManager } from './engine/engine-manager';
 import { registerAddDependencyHandler } from './maven/add-dependency';
 import { IvyProjectExplorer } from './project-explorer/ivy-project-explorer';
-import { NewProjectTool } from './project-explorer/new-project';
 import { resolveExtensionVersion } from './version/extension-version';
 import { showRuntimeLog } from './views/runtimelog-view';
 
@@ -34,7 +33,7 @@ export async function activate(context: ExtensionContext): Promise<MessengerDiag
   registerCommand('ivyPanelView.openWelcomePage', context, () => showWelcomePage(context));
   registerCommand('ivy.showStatusBarQuickPick', context, () => showStatusBarQuickPick());
 
-  registerTool('new_axon_ivy_project', context, new NewProjectTool());
+  registerTools(context);
 
   IvyDiagnostics.init(context);
   setStatusBarIcon();
